fixed include path errors, fixed warnings

+++ inc/nxp/LPC8XX.hpp/LPC81X_iocon.hpp
@@ -36,7 +36,8 @@ struct iocon {
   template <typename T>
   void setup(T &pin, pullModes mode, pinFiltering filter, clockDivider clock, uint32_t options) {
     static_assert(pin.typeFlags == pinTypeFlags::NORMAL, "only normal pins can use this setup function");
-    regs()->PIO[pin.ioconIndex] = mode | filter | clock | options;
+    regs()->PIO[pin.ioconIndex] =
+      static_cast<uint32_t>(mode) | static_cast<uint32_t>(filter) | static_cast<uint32_t>(clock) | options;
   }
   /**
    * @brief Setup normal IOCON pin
@@ -49,7 +50,7 @@ struct iocon {
   template <typename T>
   void setup(T &pin, pullModes mode, uint32_t options) {
     static_assert(pin.typeFlags == pinTypeFlags::NORMAL, "only normal pins can use this setup function");
-    regs()->PIO[pin.ioconIndex] = mode | options;
+    regs()->PIO[pin.ioconIndex] = static_cast<uint32_t>(mode) | options;
   }
   /**
    * @brief Setup normal IOCON pin
@@ -60,7 +61,7 @@ struct iocon {
   template <typename T>
   void setup(T &pin, pullModes mode) {
     static_assert(pin.typeFlags == pinTypeFlags::NORMAL, "only normal pins can use this setup function");
-    regs()->PIO[pin.ioconIndex] = mode;
+    regs()->PIO[pin.ioconIndex] = static_cast<uint32_t>(mode);
   }
   /**
    * @brief Setup I2C IOCON pin
@@ -75,7 +76,8 @@ struct iocon {
   template <typename T>
   void setup(T &pin, i2cmodes mode, pinFiltering filter, clockDivider clock, uint32_t options) {
     static_assert(pin.typeFlags & pinTypeFlags::IOCON_I2C, "only I2C pins have a i2c mode setup");
-    regs()->PIO[pin.ioconIndex] = mode | filter | clock | options;
+    regs()->PIO[pin.ioconIndex] =
+      static_cast<uint32_t>(mode) | static_cast<uint32_t>(filter) | static_cast<uint32_t>(clock) | options;
   }
   /**
    * @brief Setup I2C IOCON pin
@@ -88,7 +90,7 @@ struct iocon {
   template <typename T>
   void setup(T &pin, i2cmodes mode, uint32_t options) {
     static_assert(pin.typeFlags & pinTypeFlags::IOCON_I2C, "only I2C pins have a i2c mode setup");
-    regs()->PIO[pin.ioconIndex] = mode | options;
+    regs()->PIO[pin.ioconIndex] = static_cast<uint32_t>(mode) | options;
   }
   /**
    * @brief Setup I2C IOCON pin
@@ -100,7 +102,7 @@ struct iocon {
   template <typename T>
   void setup(T &pin, i2cmodes mode) {
     static_assert(pin.typeFlags & pinTypeFlags::IOCON_I2C, "only I2C pins have a i2c mode setup");
-    regs()->PIO[pin.ioconIndex] = mode;
+    regs()->PIO[pin.ioconIndex] = static_cast<uint32_t>(mode);
   }
 };
 }  // namespace iocon
